+++ NEWS.md
@@ -1,7 +1,13 @@
 # workflows (development version)
 
+* Added a standalone file `standalone-input-names.R` with APIs for returning the
+names of the predictors in the original data given to `fit()`. 
+
+* Each of the `pull_*()` functions soft-deprecated in workflows v0.2.3 now warn on every usage. 
+
 * Implemented postprocessing. (TODO: fill this in.)
     - `extract_tailor()` extracts the tailor from a workflow (#301).
+
 
 * Increased the minimum required R version to R 4.1.